% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_normFactors.R
\name{apply_normFactors}
\alias{apply_normFactors}
\title{Normalize binned raw counts using normalization factors}
\usage{
apply_normFactors(
  norm_method,
  out_dir,
  genome_assembly,
  treated_sample_bedfiles,
  wildtype_sample_bedfiles,
  treated_condition_label,
  wildtype_condition_label
)
}
\arguments{
\item{norm_method}{a character string specifying the normalization method to use. Allowed values include "DESeq2", "edgeR" or "depth-norm-only". Choosing "depth-norm-only" will return counts per million without any factor normalization.}

\item{out_dir}{a character string specifying the output directory for the sizeFactor-normalized BED files.}

\item{genome_assembly}{a character string specifying the genome assembly. Allowed values include "hg38" or "mm10".}

\item{treated_sample_bedfiles}{a vector specifying the BED file(s) for the treated sample(s) with raw counts.}

\item{wildtype_sample_bedfiles}{a vector specifying the BED file(s) for the wildtype sample(s) with raw counts.}

\item{treated_condition_label}{a character string specifying the condition for the treated sample(s).}

\item{wildtype_condition_label}{a character string specifying the condition for the wildtype sample(s).}
}
\value{
BED files with normalized counts
}
\description{
Normalizes binned raw counts across samples using normalization factors, which are calculated using DESeq2's median ratio method or edgeR's trimmed mean of M-values (TMM)
}
\details{
Title
}
\examples{
\dontrun{
apply_normFactors(norm_method="DESeq2",out_dir, mm10, treated_sample_bedfiles = c("NSD1KO_rep1.bed", "NSD1KO_rep2.bed"), wildtype_sample_bedfiles = c("WT_rep1.bed", "WT_rep2.bed"), treated_condition_label = "NSD1KO", wildtype_condition_label = "WT")
}
}
