% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChIPbinner-package.R
\docType{package}
\name{ChIPbinner-package}
\alias{ChIPbinner}
\alias{ChIPbinner-package}
\title{ChIPbinner: This package is designed for analyzing binned ChIP-seq or Cut&Run data for broad histone marks}
\description{
This package will analyze binned ChIP-seq/Cut&Run data. You can visualize genome-wide changes using this software and subsequently analyze clusters of similarly-behaving genomic regions.
}
\author{
\strong{Maintainer}: Reinnier Padilla \email{reinnier.padilla@mail.mcgill.ca} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
