% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cons.R
\name{cons}
\alias{cons}
\title{Title}
\usage{
cons(
  wd,
  treatment,
  control,
  cell_line,
  histone_mark,
  gene,
  intergenic,
  cluster_size,
  number_of_clusters,
  window_size
)
}
\arguments{
\item{window_size}{}
}
\description{
Title
}
