% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_based_scatterplot.R
\name{density_based_scatterplot}
\alias{density_based_scatterplot}
\title{Generate density-based clusters.}
\usage{
density_based_scatterplot(
  out_dir,
  genome_assembly,
  treated_samp_norm_bw,
  wildtype_samp_norm_bw,
  are_R_objects = FALSE,
  cell_line,
  histone_mark,
  annotated_clusters,
  number_of_clusters,
  output_filename,
  title_of_plot,
  pow = NULL,
  show_legend = FALSE,
  min,
  max,
  bin_size,
  show_scales = TRUE,
  xaxis_label,
  yaxis_label,
  height_of_figure,
  width_of_figure
)
}
\arguments{
\item{out_dir}{Output directory for the scatterplot.}

\item{genome_assembly}{Must of be one of hg38 or mm10.}

\item{treated_samp_norm_bw}{The normalized bigwig file for the treated sample.}

\item{wildtype_samp_norm_bw}{The normalized bigwig file for the wildtype sample.}

\item{are_R_objects}{Boolean term (true or false) to indicate if the inputted bigwig files are R objects. It'll use load() for the reps as opposed to reading them in via rtracklayer::import.bed(). Default to FALSE.}

\item{cell_line}{The cell line of the samples.}

\item{histone_mark}{The broad histone mark used for the analysis.}

\item{annotated_clusters}{The annotated clusters generated using annotate_clust(), which is an R object that needs to be loaded.}

\item{number_of_clusters}{The total number of clusters identified by the HDBSCAN algorithm.}

\item{output_filename}{Filename for the resuting scatterplot.}

\item{title_of_plot}{The title of the plot.}

\item{pow}{The power of. Returns the value of x to the power of y (x^y) and this is used for the scales (i.e. show bins if surpassing a certain intensity). Defaults to 1.25.}

\item{show_legend}{Boolean term whether to show legend or not.}

\item{min}{The min for the x and y axis.}

\item{max}{The max for the x and y axis}

\item{bin_size}{The number of bins represented by each hexagonal shape.}

\item{show_scales}{Boolean term whether to show scales or not.}

\item{xaxis_label}{Label for the x-axis. This is normally the baseline label (i.e. WT)}

\item{yaxis_label}{Label for the y-axis. This is normally the treated sample label.}

\item{height_of_figure}{The height of the plot.}

\item{width_of_figure}{The width of the plot.}
}
\value{
a plot of density-based clusters.
}
\description{
This function identifies similarly-behaving genomic compartments, such as regions with the highest loss or gain of a specific histone mark.
}
\details{
Title
}
\examples{
library(tidyverse)
library(rtracklayer)
library(GenomicRanges)
library(data.table)
library(lattice)
library(gridExtra)
library(DiffBind)
library(patchwork)
library(sf)
library(MASS)
library(lwgeom)
library(ggrepel)
library(hexbin)
library(ggrastr)
library(viridis)
library(pals)
library(isoband)
# using data in /extdata folder
# system.file("extdata", "downsampled.Cal27.WT_input.H3K36me2.10kb.bed.gz", package = "ChIPbinner")
density_based_scatterplot(
  out_dir = system.file("extdata", package = "ChIPbinner"),
  genome_assembly = "hg38",
  are_R_objects = TRUE,
  output_filename = "complete",
  wildtype_samp_norm_bw = system.file("extdata/data","Cal27_WT_H3K36me2_10kb_norm_bw.rda", package = "ChIPbinner"),
  treated_samp_norm_bw = system.file("extdata/data","Cal27_NSD1_KO_H3K36me2_10kb_norm_bw.rda", package = "ChIPbinner"),
  cell_line = "Cal27",
  histone_mark = "H3K36me2",
  annotated_clusters = system.file("extdata/data","complete_annotated_clusters.rda", package = "ChIPbinner"),
  number_of_clusters = 3,
  title_of_plot = "H3K36me2",
  pow = 1.1,
  show_legend = TRUE,
  min = -5,
  max = 2,
  bin_size = 50,
  show_scales = FALSE,
  xaxis_label = "WT",
  yaxis_label = "NSD1_KO",
  height_of_figure = 6,
  width_of_figure = 15
)
# using data in /data folder
# density_based_scatterplot(
#   out_dir = system.file("extdata", package = "ChIPbinner"),
#   genome_assembly = "hg38",
#   are_R_objects = TRUE,
#   output_filename = "complete",
#   wildtype_samp_norm_bw = system.file("data/Cal27_WT_H3K36me2_10kb_norm_bw.rda", package = "ChIPbinner"),
#   treated_samp_norm_bw = system.file("data/Cal27_NSD1_KO_H3K36me2_10kb_norm_bw.rda", package = "ChIPbinner"),
#   cell_line = "Cal27",
#   histone_mark = "H3K36me2",
#   annotated_clusters = system.file("data/complete_annotated_clusters.rda", package = "ChIPbinner"),
#   number_of_clusters = 3,
#   title_of_plot = "H3K36me2",
#   pow = 1.1,
#   show_legend = TRUE,
#   min = -5,
#   max = 2,
#   bin_size = 50,
#   show_scales = FALSE,
#   xaxis_label = "WT",
#   yaxis_label = "NSD1_KO",
#   height_of_figure = 6,
#   width_of_figure = 15
# )
}
