% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_based_scatterplots.R
\name{density_based_scatterplots}
\alias{density_based_scatterplots}
\title{Title}
\usage{
density_based_scatterplots(
  wd,
  path_for_norm_bw,
  treatment,
  control,
  cell_line,
  histone_mark,
  gene,
  intergenic,
  title_of_plot,
  window_size,
  number_of_clusters,
  pow = NULL,
  show_legend = FALSE,
  min,
  max,
  bin_size,
  show_scales = TRUE,
  xaxis_label,
  yaxis_label,
  height_of_figure,
  width_of_figure
)
}
\arguments{
\item{width_of_figure}{}
}
\description{
Title
}
