% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_low_counts.R
\name{filter_low_counts}
\alias{filter_low_counts}
\title{filter bins depleted in signal across all samples}
\usage{
filter_low_counts(out_dir, genome_assembly, sample_bedfiles, cutoff = 100)
}
\arguments{
\item{out_dir}{a character string specifying the output directory for the sizeFactor-normalized BED files.}

\item{genome_assembly}{a character string specifying the genome assembly. Allowed values include "hg38" or "mm10".}

\item{sample_bedfiles}{a vector specifying the BED file(s) for the samples with raw counts.}

\item{cutoff}{a numeric value specifying the raw read count cutoff for bins. Defaults to 100.}
}
\value{
BED files with filtered counts
}
\description{
Filters bins with raw read count consistently lower than a specified value across all samples
}
\details{
Title
}
\examples{
\dontrun{
filter_low_counts(out_dir, hg38, sample_bedfiles = c("NSD1KO_rep1.bed", "NSD1KO_rep2.bed","WT_rep1.bed", "WT_rep2.bed"),cutoff=100)
}
}
