% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_clust.R
\name{generate_clust}
\alias{generate_clust}
\title{Computation of density-based clusters based on the HDBSCAN algorithm.}
\usage{
generate_clust(output_file_name, out_dir, matrix_file, minpts, minsamps, cores)
}
\arguments{
\item{output_file_name}{a character string specifying the output file name.}

\item{out_dir}{a character string specifying the output directory.}

\item{matrix_file}{a character string specifying the matrix file of enrichment scores for the two samples being compared, generated from the pre_clus() function.}

\item{minpts}{an integer specifying the minimum number of data points for a cluster size. Set it to the smallest size grouping that you wish to consider a cluster.}

\item{minsamps}{an integer indicating how conservative you want your clustering to be. The larger the value of 'min_samps' you provide, the more conservative the clustering – more points will be declared as noise, and clusters will be restricted to progressively more dense areas.}

\item{cores}{an integer specifying the number of parallel cores to be used.}
}
\value{
A text (.txt) output designating each bin from the matrix file as belonging to a specific cluster or not.
}
\description{
Computes density-based clusters needed to generate the density-based scatterplots and for other downstream analysis within the package. Python version 3.9 required to run HDBSCAN, which is the underlying software used to identify clusters of bins.
}
\details{
Title
}
\examples{
generate_clust(
  output_file_name = "downsampled.Cal27.WT.NSD1_KO.H3K36me2.10kb",
  out_dir = system.file("extdata",package = "ChIPbinner"),
  matrix_file = "downsampled.Cal27.WT.NSD1_KO.H3K36me2.10kb_mat.csv",
  minpts = 100,
  minsamps = 100,
  cores = 1
)
}
