% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChIPbinner.R
\name{generate_norm_bw}
\alias{generate_norm_bw}
\title{Generate normalized bigwig.}
\usage{
generate_norm_bw(
  wd,
  chromSizes,
  binned_files_path,
  treatment,
  control,
  table,
  table_samp_name,
  cell_line,
  histone_mark,
  blacklist,
  addition,
  raw_count_cutoff,
  use_input
)
}
\arguments{
\item{wd}{working directory.}

\item{chromSizes}{chromosome sizes.}

\item{binned_files_path}{directory for binned files. Files should be in the format $line.$samp.$mark.10kb.bed.}

\item{treatment}{treatment file name.}

\item{control}{control file name,}

\item{table}{table for sequencing depth.}

\item{table_samp_name}{name in sequencing depth table.}

\item{cell_line}{cell line.}

\item{histone_mark}{histone mark.}

\item{blacklist}{blacklist file.}

\item{addition}{add.}

\item{raw_count_cutoff}{raw cutoff.}

\item{use_input}{whether to include input in calculating normalized bigwigs or not.}
}
\value{
normalized bigwig files.
}
\description{
Generate normalized bigwig.
}
\examples{
generate_norm_bw(ChIP_files)
}
