% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genic_intergenic_scatterplot.R
\name{genic_intergenic_scatterplot}
\alias{genic_intergenic_scatterplot}
\title{Title}
\usage{
genic_intergenic_scatterplot(
  path_for_norm_bw,
  out_dir,
  gene,
  intergenic,
  cell_line,
  histone_mark,
  treated_samp_label,
  control_label,
  window_size,
  title_of_plot,
  max_x,
  max_y,
  pow = NULL,
  xaxis_label,
  yaxis_label,
  min_x = NULL,
  min_y = NULL,
  show_scales = TRUE
)
}
\arguments{
\item{show_scales}{}
}
\description{
Title
}
