% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genic_intergenic_scatterplots.R
\name{genic_intergenic_scatterplots}
\alias{genic_intergenic_scatterplots}
\title{Title}
\usage{
genic_intergenic_scatterplots(
  wd,
  path,
  gene,
  intergenic,
  cell_line,
  treatment,
  control,
  histone_mark,
  window_size,
  title_of_plot,
  max_x,
  max_y,
  pow = NULL,
  xaxis_label,
  yaxis_label,
  min_x = NULL,
  min_y = NULL,
  show_scales = TRUE
)
}
\arguments{
\item{show_scales}{}
}
\description{
Title
}
