% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_norm_bw.R
\name{merge_norm_bw}
\alias{merge_norm_bw}
\title{Merge normalized bigWig files.}
\usage{
merge_norm_bw(
  path_for_norm_bw,
  rep1,
  rep2,
  merged_sample_name,
  cell_line,
  histone_mark,
  window_size,
  out_dir
)
}
\arguments{
\item{path_for_norm_bw}{Directory where the normalized bigwigs generated by 'norm.bw()' is found.}

\item{rep1}{Replicate 1 to be merged.}

\item{rep2}{Replicate 2 to be merged.}

\item{merged_sample_name}{The sample name of the merged bigWig file.}

\item{cell_line}{The cell line for the merged bigWig file.}

\item{histone_mark}{The broad histone mark used for the analysis.}

\item{window_size}{The window size in kilobytes. For example, use 10 for samples binned in 10 kilobyte (kb) bins.}

\item{out_dir}{The output directory for merged bigWig files.}
}
\value{
a single merged bigWig file.
}
\description{
This function will merge normalized bigWig files, which are normally replicates. This function should be used after implementing the 'norm_bw()' function.
}
\details{
Title
}
