% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_correlation.R
\name{plot_correlation}
\alias{plot_correlation}
\title{Generates correlation plot.}
\usage{
plot_correlation(
  ...,
  out_dir,
  output_filename,
  sample_labels = NULL,
  correlation_method = "pearson",
  colors,
  plot_title = NULL,
  key_title = NULL,
  font_size = NULL,
  col_label_angle = NULL,
  row_label_angle = NULL,
  row_adj_label = NULL,
  col_adj_label = NULL,
  plot_width = NULL,
  plot_height = NULL
)
}
\arguments{
\item{...}{file paths to normalized bigWig files. Users can list as many files as needed for the samples to be plotted.}

\item{out_dir}{a character string specifying the output directory for the scatterplot.}

\item{output_filename}{a character string specifying the file name for the PCA plot to be saved on disk.}

\item{sample_labels}{a vector of sample labels corresponding to the inputted files. If not supplied, defaults to the input filenames. The length should match the number of files.}

\item{correlation_method}{a character string specifying which correlation coefficient is to be computed. Must be one of "pearson" (default) or "spearman".}

\item{colors}{a vector of colors corresponding to each file. The length should match the number of files.}

\item{plot_title}{a character string specifying the title of the plot.}

\item{key_title}{a character string specifying the legend title.}

\item{font_size}{an integer specifying the font sizes for the text in the plot.}

\item{col_label_angle}{an integer specifying the angle of the labels for the columns. If not supplied, defaults to 360.}

\item{row_label_angle}{an integer specifying the angle of the labels for the rows. If not supplied, defaults to 270.}

\item{row_adj_label}{a 2-element vector giving the (left-right, top-bottom) justification of row labels (relative to the text orientation). Defaults to c(0.5,0.5).}

\item{col_adj_label}{a 2-element vector giving the (left-right, top-bottom) justification of column labels (relative to the text orientation). Defaults to c(0.5,0.5).}

\item{plot_width}{a numeric specifying the width of the scatterplot. If not supplied, defaults to 8.}

\item{plot_height}{a numeric specifying the height of the scatterplot. If not supplied, defaults to 10.}
}
\value{
a correlation plot
}
\description{
Plots a correlation matrix-based hierarchical clustering.
}
\details{
Title
}
\examples{
\dontrun{
file1 <- "file1.bw"
file2 <- "file2.bw"
colors <- c("red", "blue")
sample_labels <- c("samp1","samp2")
result <- plot_correlation(file1, file2, sample_labels = sample_labels,colors = colors,out_dir=outdir,output_filename=output,plot_title=title)
}
}
