% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_PCA.R
\name{plot_PCA}
\alias{plot_PCA}
\title{Plot PCA.}
\usage{
plot_PCA(
  ...,
  out_dir,
  output_filename,
  colors,
  plot_title,
  sample_labels = NULL,
  font_size = NULL,
  point_size = NULL,
  plot_width = NULL,
  plot_height = NULL
)
}
\arguments{
\item{...}{file paths to normalized bigWig files. Users can list as many files as needed for the samples to be plotted.}

\item{out_dir}{a character string specifying the output directory for the scatterplot.}

\item{output_filename}{a character string specifying the file name for the PCA plot to be saved on disk.}

\item{colors}{a vector of colors corresponding to each file. The length should match the number of files.}

\item{plot_title}{a character string specifying the title of the plot.}

\item{sample_labels}{a vector of sample labels corresponding to the inputted files. If not supplied, defaults to the input filenames. The length should match the number of files.}

\item{font_size}{an integer specifying the font sizes for the text in the plot.}

\item{point_size}{an integer specifying the dot size in the plot.}

\item{plot_width}{a numeric specifying the width of the scatterplot. If not supplied, defaults to 5.}

\item{plot_height}{a numeric specifying the height of the scatterplot. If not supplied, defaults to 5.}
}
\value{
a PCA plot
}
\description{
Plots the results from principal component analysis using samples that have been normalized.
}
\details{
Title
}
\examples{
\dontrun{
file1 <- "file1.bw"
file2 <- "file2.bw"
colors <- c("red", "blue")
sample_labels <- c("samp1","samp2")
result <- plotPCA(file1, file2, sample_labels=sample_labels, colors = colors,out_dir=outdir,output_filename=output,plot_title=title)
}
}
