% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_clus.R
\name{pre_clus}
\alias{pre_clus}
\title{Pre-processing the normalized bigWig files prior to generating clusters.}
\usage{
pre_clus(
  path_for_norm_bw,
  out_dir,
  window_size,
  treated_samp_label,
  baseline_samp_label,
  cell_line,
  histone_mark
)
}
\arguments{
\item{path_for_norm_bw}{Directory where the normalized bigwigs generated by 'norm.bw()' is found.}

\item{out_dir}{Output directory for the matrix, BED file and preliminary scatterplot.}

\item{window_size}{The window size in kilobytes - this number must be an integer. For example, use 10 for samples binned in 10 kilobyte (kb) bins.}

\item{treated_samp_label}{The treated sample name/label.}

\item{baseline_samp_label}{The baseline sample name/label to which to compare the treated sample. This is usually the wildtype (WT) sample.}

\item{cell_line}{The cell line of the samples.}

\item{histone_mark}{The broad histone mark used for the analysis.}
}
\value{
a matrix file of scores for the two samples being compared, a pooled BED file of genomic coordinates for the bins and a preliminary scatterplot of the comparison.
}
\description{
This function pre-processes the bigWig files: it finds overlapping regions between two bigWig files, removes the bottom and top 1\% of bins across the two samples and finally generates a matrix of scores and BED file of coordinates to be used with HDBscan for generating clusters of bins with similar scores.
}
\details{
Title
}
\examples{
library(tidyverse)
library(rtracklayer)
library(GenomicRanges)
library(data.table)
library(lattice)
library(gridExtra)
library(DiffBind)
library(idr2d)
library(patchwork)
pre_clus(path_for_norm_bw = system.file("extdata/norm_bw", package = "ChIPbinner"),
         out_dir = system.file("extdata", package = "ChIPbinner"),
         window_size = 10,
         treated_samp_label = "NSD1_KO",
         baseline_samp_label = "WT",
         cell_line = "Cal27",
         histone_mark = "H3K36me2")
}
