% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_clust.R
\name{pre_clust}
\alias{pre_clust}
\title{Pre-processing of normalized bigWig files.}
\usage{
pre_clust(
  out_dir,
  treated_samp_norm_bw,
  wildtype_samp_norm_bw,
  output_filename,
  are_R_objects = FALSE
)
}
\arguments{
\item{out_dir}{a character string specifying the output directory for the matrix of scores, BED file of genomic coordinates and preliminary density-based scatterplot.}

\item{treated_samp_norm_bw}{a character string specifying the normalized bigWig file for the treated sample.}

\item{wildtype_samp_norm_bw}{a character string specifying the normalized bigWig file for the wildtype sample.}

\item{output_filename}{a character string specifying the file-name for the matrix and pooled BED file.}

\item{are_R_objects}{a logical indicating whether the inputted bigwig files are R objects. It'll use load() for the reps as opposed to reading them in via rtracklayer::import.bed(). Defaults to FALSE.}
}
\value{
a matrix file of scores, a pooled BED file of genomic coordinates for the bins and a preliminary density-based scatterplot for the two samples being compared.
}
\description{
Finds overlapping regions between two bigWig files, removes the bottom and top 1\% of bins across the two samples and finally generates a matrix of scores and BED file of coordinates to be used with HDBScan for generating clusters of bins with similar scores.
}
\details{
Title
}
\examples{
pre_clust(
  out_dir = system.file("extdata",package = "ChIPbinner"),
  treated_samp_norm_bw = system.file("extdata/data","downsampled.Cal27.WT.H3K36me2.10kb.norm_bw.rda", package = "ChIPbinner"),
  wildtype_samp_norm_bw = system.file("extdata/data","downsampled.Cal27.NSD1_KO.H3K36me2.10kb.norm_bw.rda", package = "ChIPbinner"),
  output_filename = "downsampled.Cal27.WT.NSD1_KO.H3K36me2.10kb",
  are_R_objects = TRUE
)
}
