% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChIPbinner.R
\name{pre_clustering}
\alias{pre_clustering}
\title{Pre-clustering.}
\usage{
pre_clustering(
  wd,
  path_for_norm_bw,
  typeof_file,
  treatment,
  control,
  cell_line,
  histone_mark
)
}
\arguments{
\item{wd}{working directory.}

\item{path_for_norm_bw}{path for normalized bigiwg.}

\item{typeof_file}{norm bigwig.}

\item{treatment}{treatment sample.}

\item{control}{control sample.}

\item{cell_line}{cell line.}

\item{histone_mark}{histone mark.}
}
\value{
comparison matrices.
}
\description{
Pre-clustering.
}
\examples{
treatment vs control sample.
}
