% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizeFactor_norm.R
\name{sizeFactor_norm}
\alias{sizeFactor_norm}
\title{Normalize binned raw counts using size factors}
\usage{
sizeFactor_norm(
  out_dir,
  genome_assembly,
  treated_sample_bedfiles,
  wildtype_sample_bedfiles,
  treated_condition_label,
  wildtype_condition_label
)
}
\arguments{
\item{out_dir}{a character string specifying the output directory for the sizeFactor-normalized BED files.}

\item{genome_assembly}{a character string specifying the genome assembly. Allowed values include "hg38" or "mm10".}

\item{treated_sample_bedfiles}{a vector specifying the BED file(s) for the treated sample(s) with raw counts.}

\item{wildtype_sample_bedfiles}{a vector specifying the BED file(s) for the wildtype sample(s) with raw counts.}

\item{treated_condition_label}{a character string specifying the condition for the treated sample(s).}

\item{wildtype_condition_label}{a character string specifying the condition for the wildtype sample(s).}
}
\value{
sizeFactor-normalized BED files
}
\description{
Normalizes binned raw counts across samples using size factors, which are calculated with the median of ratios method from DESeq2
}
\details{
Title
}
\examples{
\dontrun{
sizeFactor_norm(out_dir,mm10,treated_sample_bedfiles=c("NSD1KO_rep1.bed","NSD1KO_rep2.bed"),wildtype_sample_bedfiles=c("WT_rep1.bed","WT_rep2.bed"),treated_condition_label="NSD1KO",wildtype_condition_label="WT")
}
}
