% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_clust.R
\name{annotate_clust}
\alias{annotate_clust}
\title{Annotate clusters.}
\usage{
annotate_clust(
  number_of_clusters,
  matrix_file,
  pooled_bed_file,
  hdbscan_output_file,
  output_filename,
  out_dir
)
}
\arguments{
\item{number_of_clusters}{The total number of clusters identified by the HDBSCAN algorithm.}

\item{matrix_file}{The matrix file of enrichment scores for the two samples being compared, generated from the pre_clus() function.}

\item{pooled_bed_file}{The pooled BED file generated from 'pre_clus()' consisting of genomic coordinates found in both samples being compared.}

\item{hdbscan_output_file}{The HDBSCAN output which specifies whether each bin falls into a specific cluster or not.}

\item{output_filename}{Name for the matrix file and pooled BED file.}

\item{out_dir}{The output directory.}
}
\value{
a R object consisting of multiple annotated clusters, each with their set of bins.
}
\description{
Annotates clusters identified after using HDBSCAN.
}
\details{
Title
}
\examples{
#!/usr/bin/env Rscript
library(tidyverse)
library(rtracklayer)
library(GenomicRanges)
library(data.table)
library(lattice)
library(gridExtra)
library(DiffBind)
library(idr2d)
library(patchwork)
# system.file("extdata", "downsampled.Cal27.WT_input.H3K36me2.10kb.bed.gz", package = "ChIPbinner")
annotate_clust(number_of_clusters = 3,
               matrix_file = system.file("extdata", "downsampled.Cal27.WT.NSD1_KO.H3K36me2.10kb_mat.csv", package = "ChIPbinner"),
               pooled_bed_file = system.file("extdata", "downsampled.Cal27.WT.NSD1_KO.H3K36me2.10kb_pooled.bed.gz", package = "ChIPbinner"),
               hdbscan_output_file= system.file("extdata", "clus.downsampled.Cal27.WT.NSD1_KO.H3K36me2.10kb.100.100.txt", package = "ChIPbinner"),
               output_filename = "downsampled",
               out_dir=system.file("extdata",package = "ChIPbinner"))
}
