% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChIPbinner-package.R
\docType{package}
\name{ChIPbinner-package}
\alias{ChIPbinner}
\alias{ChIPbinner-package}
\title{ChIPbinner: Analyze ChIP-seq/Cut&Run data in binned windows for broad histone marks}
\description{
A system for clustering and analyzing ChIP-seq or Cut&Run data that has been binned in windows. You provide two samples for comparison, input information regarding the samples and follow the workflow. ChIPbinner will identify clusters of similarly behaving bins, assign bins and clusters of bins as genic or intergenic and finally, run enrichment and depletion analysis for the annotated clusters.
}
\author{
\strong{Maintainer}: Reinnier Padilla \email{reinnier.padilla@mail.mcgill.ca} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
