% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_based_scatterplot.R
\name{density_based_scatterplot}
\alias{density_based_scatterplot}
\title{Generate density-based clusters.}
\usage{
density_based_scatterplot(
  path_for_norm_bw,
  out_dir,
  treated_samp_label,
  baseline_samp_label,
  cell_line,
  histone_mark,
  matrix_file,
  pooled_bed_file,
  hdbscan_output_file,
  number_of_clusters,
  gene,
  intergenic,
  title_of_plot,
  window_size,
  pow = NULL,
  show_legend = FALSE,
  min,
  max,
  bin_size,
  show_scales = TRUE,
  xaxis_label,
  yaxis_label,
  height_of_figure,
  width_of_figure
)
}
\arguments{
\item{path_for_norm_bw}{Directory where the normalized bigwigs generated by 'norm.bw()' is found.}

\item{out_dir}{Output directory for the scatterplot.}

\item{treated_samp_label}{The treated sample name/label.}

\item{baseline_samp_label}{The baseline sample name/label to which to compare the treated sample. This is usually the wildtype (WT) sample.}

\item{cell_line}{The cell line of the samples.}

\item{histone_mark}{The broad histone mark used for the analysis.}

\item{matrix_file}{The matrix file generated from 'pre_clus()', consisting of only the enrichment scores per bin for each of the two samples being compared.}

\item{pooled_bed_file}{The pooled BED file generated from 'pre_clus()' consisting of genomic coordinates found in both samples being compared.}

\item{hdbscan_output_file}{The HDBSCAN output which specifies whether each bin falls into a specific cluster or not.}

\item{number_of_clusters}{The total number of clusters identified by the HDBSCAN algorithm.}

\item{gene}{Annotated genic regions in BED format. hg38 and mm10 are included in the package and can be accessed in 'extdata'.}

\item{intergenic}{Annotated intergenic regions in BED format. hg38 and mm10 are included in the package and can be accessed in 'extdata'.}

\item{title_of_plot}{The title of the plot.}

\item{window_size}{The window size in kilobytes. For example, use 10 for samples binned in 10 kilobyte (kb) bins.}

\item{pow}{The power of. Returns the value of x to the power of y (x^y) and this is used for the scales (i.e. show bins if surpassing a certain intensity). Defaults to 1.25.}

\item{show_legend}{Boolean term whether to show legend or not.}

\item{min}{The min for the x and y axis.}

\item{max}{The max for the x and y axis}

\item{bin_size}{The number of bins represented by each hexagonal shape.}

\item{show_scales}{Boolean term whether to show scales or not.}

\item{xaxis_label}{Label for the x-axis. This is normally the baseline label (i.e. WT)}

\item{yaxis_label}{Label for the y-axis. This is normally the treated sample label.}

\item{height_of_figure}{The height of the plot.}

\item{width_of_figure}{The width of the plot.}
}
\value{
a plot of density-based clusters.
}
\description{
This function identifies similarly-behaving genomic compartments, such as regions with the highest loss or gain of a specific histone mark.
}
\details{
Title
}
\examples{
library(tidyverse)
library(rtracklayer)
library(GenomicRanges)
library(data.table)
library(lattice)
library(gridExtra)
library(DiffBind)
library(patchwork)
library(sf)
library(MASS)
library(lwgeom)
library(ggrepel)
library(hexbin)
library(ggrastr)
library(viridis)
library(pals)
library(isoband)
density_based_scatterplot(path_for_norm_bw = system.file("extdata/norm_bw", package = "ChIPbinner"),
                          out_dir = system.file("extdata", package = "ChIPbinner"),
                          treated_samp_label = "NSD1_KO",
                          baseline_samp_label = "WT",
                          cell_line = "Cal27",
                          histone_mark = "H3K36me2",
                          matrix_file = system.file("extdata","Cal27.WT.NSD1_KO.H3K36me2.10kb.mat.csv", package = "ChIPbinner"),
                          pooled_bed_file = system.file("extdata","Cal27.WT.NSD1_KO.H3K36me2.10kb.pooled.bed", package = "ChIPbinner"),
                          hdbscan_output_file = system.file("extdata","clus.Cal27.WT.NSD1_KO.H3K36me2.10kb.5000.5000.txt", package = "ChIPbinner"),
                          number_of_clusters = 3,
                          gene = system.file("extdata", "hg38_gene.bed", package = "ChIPbinner"),
                          intergenic = system.file("extdata", "hg38_intergenic.bed", package = "ChIPbinner"),
                          title_of_plot = "H3K36me2",
                          window_size = 10,
                          pow = 1.25,
                          show_legend = TRUE,
                          min=-5,
                          max=2,
                          bin_size=50,
                          show_scales = FALSE,
                          xaxis_label="WT",
                          yaxis_label="NSD1_KO",
                          height_of_figure=6,
                          width_of_figure=15)
}
