% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_based_scatterplot.R
\name{density_based_scatterplot}
\alias{density_based_scatterplot}
\title{Plot scatterplots of annotated genic/intergenic regions and density-based clusters.}
\usage{
density_based_scatterplot(
  out_dir,
  genome_assembly,
  treated_samp_norm_bw,
  wildtype_samp_norm_bw,
  are_R_objects = FALSE,
  cell_line,
  histone_mark,
  annotated_clusters,
  number_of_clusters,
  output_filename,
  title_of_plot,
  plot_title_font_size = 12,
  pow = NULL,
  show_legend = FALSE,
  legend_font_size = 7,
  min,
  max,
  bin_size,
  show_scales = TRUE,
  xaxis_label,
  yaxis_label,
  axis_title_font_size = 10,
  height_of_figure = 6,
  width_of_figure = 15
)
}
\arguments{
\item{out_dir}{a character string specifying the output directory for the scatterplot.}

\item{genome_assembly}{a character string specifying the genome assembly. Allowed values include "hg38" or "mm10".}

\item{treated_samp_norm_bw}{a character string specifying the normalized bigwig file for the treated sample.}

\item{wildtype_samp_norm_bw}{a character string specifying the normalized bigwig file for the wildtype sample.}

\item{are_R_objects}{a logical indicating whether the inputted bigwig files are R objects. It'll use load() for the reps as opposed to reading them in via rtracklayer::import.bed(). Defaults to FALSE.}

\item{cell_line}{a character string specifying the cell line of the samples.}

\item{histone_mark}{a character string specifying the broad histone mark being analyzed.}

\item{annotated_clusters}{a character string specifying the R object containing the annotated clusters generated using annotate_clust().}

\item{number_of_clusters}{an integer specifying the total number of clusters identified by the HDBSCAN algorithm.}

\item{output_filename}{a character string specifying the file name for the resuting scatterplots to be saved on disk.}

\item{title_of_plot}{a character string specifying title of the plot.}

\item{plot_title_font_size}{An integer specifying the font size for each plot title. If not supplied, defaults to 12.}

\item{pow}{a numeric specifying the power of. Returns the value of x to the power of y (x^y) and this is used for the scales (i.e. show bins if surpassing a certain intensity). Defaults to 1.25.}

\item{show_legend}{a logical indicating whether to show legend or not. Defaults to false.}

\item{legend_font_size}{An integer specifying the font size for the legends. If not supplied, defaults to 7.}

\item{min}{a numeric specifying the minimum value for the x and y axis.}

\item{max}{a numeric specifying the maximum value for the x and y axis}

\item{bin_size}{an integer specifying the number of bins represented by each hexagonal shape.}

\item{show_scales}{a logical indicating whether to show scales or not. Defaults to TRUE.}

\item{xaxis_label}{a character string specifying the label for the x-axis. This is normally the wildtype sample.}

\item{yaxis_label}{a character string specifying the label for the y-axis. This is normally the treated sample.}

\item{axis_title_font_size}{An integer specifying the font size for the axis titles. If not supplied, defaults to 10.}

\item{height_of_figure}{A numeric specifying the height of the plots. If not supplied, defaults to 5.}

\item{width_of_figure}{A numeric specifying the width of the plots. If not supplied, defaults to 5.}
}
\value{
plots of annotated genic/intergenic bins and previously identified density-based clusters.
}
\description{
Plots scatterplots of annotated genic/intergenic regions and previously identified clusters of similar-behaving genomic regions.
}
\details{
Title
}
\examples{
density_based_scatterplot(
  out_dir = system.file("extdata", package = "ChIPbinner"),
  genome_assembly = "hg38",
  are_R_objects = TRUE,
  output_filename = "complete",
  wildtype_samp_norm_bw = system.file("extdata/data","Cal27_WT_H3K36me2_10kb_norm_bw.rda", package = "ChIPbinner"),
  treated_samp_norm_bw = system.file("extdata/data","Cal27_NSD1_KO_H3K36me2_10kb_norm_bw.rda", package = "ChIPbinner"),
  cell_line = "Cal27",
  histone_mark = "H3K36me2",
  annotated_clusters = system.file("extdata/data","complete_annotated_clusters.rda", package = "ChIPbinner"),
  number_of_clusters = 3,
  title_of_plot = "H3K36me2",
  pow = 1.1,
  show_legend = TRUE,
  min = -5,
  max = 2,
  bin_size = 50,
  show_scales = FALSE,
  xaxis_label = "WT",
  yaxis_label = "NSD1_KO",
  height_of_figure = 6,
  width_of_figure = 15
)
# using data in /data folder
# density_based_scatterplot(
#   out_dir = system.file("extdata", package = "ChIPbinner"),
#   genome_assembly = "hg38",
#   are_R_objects = TRUE,
#   output_filename = "complete",
#   wildtype_samp_norm_bw = system.file("data/Cal27_WT_H3K36me2_10kb_norm_bw.rda", package = "ChIPbinner"),
#   treated_samp_norm_bw = system.file("data/Cal27_NSD1_KO_H3K36me2_10kb_norm_bw.rda", package = "ChIPbinner"),
#   cell_line = "Cal27",
#   histone_mark = "H3K36me2",
#   annotated_clusters = system.file("data/complete_annotated_clusters.rda", package = "ChIPbinner"),
#   number_of_clusters = 3,
#   title_of_plot = "H3K36me2",
#   pow = 1.1,
#   show_legend = TRUE,
#   min = -5,
#   max = 2,
#   bin_size = 50,
#   show_scales = FALSE,
#   xaxis_label = "WT",
#   yaxis_label = "NSD1_KO",
#   height_of_figure = 6,
#   width_of_figure = 15
# )
}
