% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_clust.R
\name{enrich_clust}
\alias{enrich_clust}
\title{Enrichment & depletion analysis for clusters of bins.}
\usage{
enrich_clust(
  genome_assembly,
  annotated_clusters,
  query_cluster,
  pooled_bed_file,
  functional_db,
  region,
  cores,
  n_elements,
  cutoff_for_overlap,
  file_plot_name,
  output_table_name,
  width_of_plot,
  height_of_plot,
  out_dir
)
}
\arguments{
\item{genome_assembly}{A character string specifying the genome assembly. Allowed values include "hg38" or "mm10".}

\item{annotated_clusters}{An R object containing annotated clusters, generated from the annotate_clust() function.}

\item{query_cluster}{A character string specifying the cluster being to be tested. Normally, the clusters are label alphabetically from A to Z. For example, "B" to specify cluster B.}

\item{pooled_bed_file}{A BED file generated from 'pre_clus()' consisting of pooled genomic coordinates found in both samples being compared. Used for the background.}

\item{functional_db}{An R object consisting of a curated database of functional annotations.}

\item{region}{A string specifying the region to be tested. Must be one of "genome_wide", "genic" or "intergenic". Genome_wide indicates global testing of bins, including those found in both intergenic and genic regions. By specifying the region to be either genic or intergenic, the user can evaluate exclusively genic or intergenic bins overlapping a specific class of annotated regions. In these cases, the background is stratified to only genic or intergenic regions to avoid spurious associations to annotations confounded by their predominantly genic or intergenic localization.}

\item{cores}{An integer specifying the number of parallel cores to be used.}

\item{n_elements}{An integer specifying the number of elements to be plotted. This will not affect the output table.}

\item{cutoff_for_overlap}{An integer specifying the minimum number of overlap between the bins in the specified cluster and the class of annotated regions.}

\item{file_plot_name}{A character string specifying the filename of the output plot.}

\item{output_table_name}{A character string specifying the filename for the output table.}

\item{width_of_plot}{A numeric specifying the width of the plot.}

\item{height_of_plot}{A numeric specifying the height of the plot.}

\item{out_dir}{A character string specifying the output directory.}
}
\value{
A table and plot of the results of the enrichment & depletion analysis.
}
\description{
Performs overlap enrichment & depletion analysis between bins found in a specific cluster and a class of annotated regions from a curated database. Uses Fisher's exact test to assess statistical significance of the overlap against a background of all bins found in both cell lines being compared.
}
\details{
Title
}
\examples{
#!/usr/bin/env Rscript
library(tidyverse)
library(readr)
library(rtracklayer)
library(GenomicRanges)
library(LOLA)
library(highcharter)
enrich_clust(
  genome_assembly = "hg38",
  annotated_clusters = system.file("extdata/data","downsampled_annotated_clusters.rda", package = "ChIPbinner"),
  query_cluster = "B",
  pooled_bed_file = system.file("extdata", "downsampled.Cal27.WT.NSD1_KO.H3K36me2.10kb_pooled.bed.gz", package = "ChIPbinner"),
  functional_db = system.file("extdata/data","downsampled_hg38_ccreDB.rda", package = "ChIPbinner"),
  region = "genome_wide",
  cores = 1,
  n_elements = 7,
  cutoff_for_overlap = 10,
  file_plot_name = "downsampled.ccre.genome_wide.enrichment_depletion",
  output_table_name = "downsampled.ccre.genome_wide.enrichment_depletion",
  width_of_plot = 7,
  height_of_plot = 3,
  out_dir = system.file("extdata", package = "ChIPbinner")
)
}
