% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genic_intergenic_scatterplot.R
\name{genic_intergenic_scatterplot}
\alias{genic_intergenic_scatterplot}
\title{Plot a scatterplot of bins annotated as genic and intergenic.}
\usage{
genic_intergenic_scatterplot(
  out_dir,
  genome_assembly,
  cell_line,
  histone_mark,
  treated_samp_norm_bw,
  wildtype_samp_norm_bw,
  are_R_objects = FALSE,
  output_filename,
  title_of_plot,
  max_x,
  max_y,
  pow = NULL,
  xaxis_label,
  yaxis_label,
  min_x = NULL,
  min_y = NULL,
  show_scales = TRUE,
  show_legend = FALSE,
  legend_pos = NULL,
  height_of_plot = NULL,
  width_of_plot = NULL
)
}
\arguments{
\item{out_dir}{a character string specifying the output directory for the scatterplot.}

\item{genome_assembly}{a character string specifying the genome assembly. Allowed values include "hg38" or "mm10".}

\item{cell_line}{a character string specifying the cell line of the samples.}

\item{histone_mark}{a character string specifying the broad histone mark being analyzed.}

\item{treated_samp_norm_bw}{a character string specifying the normalized bigwig file for the treated sample.}

\item{wildtype_samp_norm_bw}{a character string specifying the normalized bigwig file for the wildtype sample.}

\item{are_R_objects}{a logical indicating whether the inputted bigwig files are R objects. It'll use load() for the reps as opposed to reading them in via rtracklayer::import.bed(). Defaults to FALSE.}

\item{output_filename}{a character string specifying the file name for the resulting scatterplot to be saved on disk.}

\item{title_of_plot}{a character string specifying title of the plot.}

\item{max_x}{a numeric specifying the maximum value for the x-axis.}

\item{max_y}{a numeric specifying the maximum value for the y-axis.}

\item{pow}{a numeric specifying the power of. Returns the value of x to the power of y (x^y) and this is used for the scales (i.e. show bins if surpassing a certain intensity). Defaults to 1.25.}

\item{xaxis_label}{a character string specifying the label for the x-axis. This is normally the wildtype sample.}

\item{yaxis_label}{a character string specifying the label for the y-axis. This is normally the treated sample.}

\item{min_x}{a numeric specifying the minimum value for the x-axis.}

\item{min_y}{a numeric specifying the minimum value for the y-axis.}

\item{show_scales}{a logical indicating whether to show scales or not. Defaults to TRUE.}

\item{show_legend}{a logical indicating whether to show legend or not. Defaults to false.}

\item{legend_pos}{a character string specifying the legend position. Allowed values include "left" or "right". Defaults to "left".}

\item{height_of_plot}{a numeric specifying the height of the scatterplot. If not supplied, defaults to 5.}

\item{width_of_plot}{a numeric specifying the width of the scatterplot. If not supplied, defaults to 5.}
}
\value{
Returns a scatterplot of bins annotated as genic or intergenic comparing two samples.
}
\description{
Using two normalized bigWig files generated using 'norm.bw', the user can generate a scatterplot to compare bins across two conditions (treated sample versus baseline sample (such as wildtype)). The bins will be annotated according to their overlap with genic and intergenic regions.
}
\details{
Title
}
\examples{
genic_intergenic_scatterplot(
  out_dir = system.file("extdata", package = "ChIPbinner"),
  genome_assembly = "hg38",
  cell_line = "Cal27",
  wildtype_samp_norm_bw = system.file("extdata/data","downsampled.Cal27.WT.H3K36me2.10kb.norm_bw.rda", package = "ChIPbinner"),
  treated_samp_norm_bw = system.file("extdata/data","downsampled.Cal27.NSD1_KO.H3K36me2.10kb.norm_bw.rda", package = "ChIPbinner"),
  are_R_objects = TRUE,
  histone_mark = "H3K36me2",
  output_filename = "downsampled_Cal27.WT_NSD1KO.H3K36me2.10kb",
  title_of_plot = "Cal27 ChIP-seq H3K36me2",
  xaxis_label = "WT",
  yaxis_label = "NSD1_KO",
  max_x = 1,
  max_y = 1,
  min_x = -5,
  min_y = -5,
  pow = 1.25,
  show_scales = FALSE,
  show_legend = TRUE,
  legend_pos = "left"
)
}
