% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genic_intergenic_scatterplot.R
\name{genic_intergenic_scatterplot}
\alias{genic_intergenic_scatterplot}
\title{Generate a scatterplot of bins annotated with genic and intergenic regions.}
\usage{
genic_intergenic_scatterplot(
  path_for_norm_bw,
  out_dir,
  gene,
  intergenic,
  cell_line,
  histone_mark,
  treated_samp_label,
  baseline_samp_label,
  window_size,
  title_of_plot,
  max_x,
  max_y,
  pow = NULL,
  xaxis_label,
  yaxis_label,
  min_x = NULL,
  min_y = NULL,
  show_scales = TRUE,
  show_legend = FALSE,
  legend_pos = NULL,
  height_of_plot = NULL,
  width_of_plot = NULL
)
}
\arguments{
\item{path_for_norm_bw}{Directory where the normalized bigwigs generated by 'norm.bw()' is found.}

\item{out_dir}{Output directory for the scatterplot.}

\item{gene}{Annotated genic regions in BED format. hg38 and mm10 are included in the package and can be accessed in 'extdata'.}

\item{intergenic}{Annotated intergenic regions in BED format. hg38 and mm10 are included in the package and can be accessed in 'extdata'.}

\item{cell_line}{The cell line of the samples.}

\item{histone_mark}{The broad histone mark used for the analysis.}

\item{treated_samp_label}{The treated sample name/label.}

\item{baseline_samp_label}{The baseline sample name/label to which to compare the treated sample. This is usually the wildtype (WT) sample.}

\item{window_size}{The window size in kilobytes. For example, use 10 for samples binned in 10 kilobyte (kb) bins.}

\item{title_of_plot}{The title of the plot.}

\item{max_x}{The maximum value for the x-axis.}

\item{max_y}{The maximum value for the y-axis.}

\item{pow}{The power of. Returns the value of x to the power of y (x^y) and this is used for the scales (i.e. show bins if surpassing a certain intensity). Defaults to 1.25.}

\item{xaxis_label}{Label for the x-axis. This is normally the baseline label (i.e. WT)}

\item{yaxis_label}{Label for the y-axis. This is normally the treated sample label.}

\item{min_x}{The minimum value for the x-axis.}

\item{min_y}{The minimum value for the y-axis.}

\item{show_scales}{Boolean term whether to show scales or not.}

\item{show_legend}{Boolean term whether to show legend or not.}

\item{legend_pos}{Legend position. The only two options are "left" or "right". Defaults to "left".}

\item{height_of_plot}{The height of the scatterplot.}

\item{width_of_plot}{The width of the scatterplot.}
}
\value{
Returns a scatterplot of bins annotated genic or intergenic comparing across two conditions.
}
\description{
Using two normalized bigWig files generated using 'norm.bw', the user can generate a scatterplot to compare bins across two conditions (treated sample versus baseline sample (such as wildtype)). The bins will be annotated according to their overlap with genic and intergenic regions.
}
\details{
Title
}
