% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genic_intergenic_scatterplot.R
\name{genic_intergenic_scatterplot}
\alias{genic_intergenic_scatterplot}
\title{Generate a scatterplot of bins annotated with genic and intergenic regions.}
\usage{
genic_intergenic_scatterplot(
  out_dir,
  genome_assembly,
  cell_line,
  histone_mark,
  treated_samp_norm_bw,
  wildtype_samp_norm_bw,
  are_R_objects = FALSE,
  output_filename,
  title_of_plot,
  max_x,
  max_y,
  pow = NULL,
  xaxis_label,
  yaxis_label,
  min_x = NULL,
  min_y = NULL,
  show_scales = TRUE,
  show_legend = FALSE,
  legend_pos = NULL,
  height_of_plot = NULL,
  width_of_plot = NULL
)
}
\arguments{
\item{out_dir}{Output directory for the scatterplot.}

\item{genome_assembly}{Must of be one of hg38 or mm10.}

\item{cell_line}{The cell line of the samples.}

\item{histone_mark}{The broad histone mark used for the analysis.}

\item{treated_samp_norm_bw}{The normalized bigwig file for the treated sample.}

\item{wildtype_samp_norm_bw}{The normalized bigwig file for the wildtype sample.}

\item{are_R_objects}{Boolean term (true or false) to indicate if the inputted bigwig files are R objects. It'll use load() for the reps as opposed to reading them in via rtracklayer::import.bed(). Default to FALSE.}

\item{output_filename}{Filename for the resuting scatterplot.}

\item{title_of_plot}{The title of the plot.}

\item{max_x}{The maximum value for the x-axis.}

\item{max_y}{The maximum value for the y-axis.}

\item{pow}{The power of. Returns the value of x to the power of y (x^y) and this is used for the scales (i.e. show bins if surpassing a certain intensity). Defaults to 1.25.}

\item{xaxis_label}{Label for the x-axis. This is normally the baseline label (i.e. WT)}

\item{yaxis_label}{Label for the y-axis. This is normally the treated sample label.}

\item{min_x}{The minimum value for the x-axis.}

\item{min_y}{The minimum value for the y-axis.}

\item{show_scales}{Boolean term whether to show scales or not.}

\item{show_legend}{Boolean term whether to show legend or not.}

\item{legend_pos}{Legend position. The only two options are "left" or "right". Defaults to "left".}

\item{height_of_plot}{The height of the scatterplot.}

\item{width_of_plot}{The width of the scatterplot.}
}
\value{
Returns a scatterplot of bins annotated genic or intergenic comparing across two conditions.
}
\description{
Using two normalized bigWig files generated using 'norm.bw', the user can generate a scatterplot to compare bins across two conditions (treated sample versus baseline sample (such as wildtype)). The bins will be annotated according to their overlap with genic and intergenic regions.
}
\details{
Title
}
\examples{
library(data.table)
library(tidyverse)
library(matrixStats)
library(isoband)
library(sf)
library(ggrepel)
library(viridis)
library(MASS)
library(lwgeom)
library(hexbin)
library(pals)
library(patchwork)
library(GenomicRanges)
library(rtracklayer)
genic_intergenic_scatterplot(
  out_dir = system.file("extdata", package = "ChIPbinner"),
  genome_assembly = "hg38",
  cell_line = "Cal27",
  wildtype_samp_norm_bw = system.file("extdata/data","downsampled.Cal27.WT.H3K36me2.10kb.norm_bw.rda", package = "ChIPbinner"),
  treated_samp_norm_bw = system.file("extdata/data","downsampled.Cal27.NSD1_KO.H3K36me2.10kb.norm_bw.rda", package = "ChIPbinner"),
  are_R_objects = TRUE,
  histone_mark = "H3K36me2",
  output_filename = "downsampled_Cal27.WT_NSD1KO.H3K36me2.10kb",
  title_of_plot = "Cal27 ChIP-seq H3K36me2",
  xaxis_label = "WT",
  yaxis_label = "NSD1_KO",
  max_x = 1,
  max_y = 1,
  min_x = -5,
  min_y = -5,
  pow = 1.25,
  show_scales = FALSE,
  show_legend = TRUE,
  legend_pos = "left"
)
}
