% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_norm_bw.R
\name{merge_norm_bw}
\alias{merge_norm_bw}
\title{Title}
\usage{
merge_norm_bw(
  wd,
  path,
  rep1,
  rep2,
  merged_sample_name,
  cell_line,
  histone_mark,
  window_size
)
}
\arguments{
\item{window_size}{}
}
\description{
Title
}
