% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_norm_bw.R
\name{merge_norm_bw}
\alias{merge_norm_bw}
\title{Merge normalized bigWig files.}
\usage{
merge_norm_bw(rep1, rep2, are_R_objects = FALSE, merged_samples_label, out_dir)
}
\arguments{
\item{rep1}{Replicate 1 to be merged. can be a bigWig file or an R object.}

\item{rep2}{Replicate 2 to be merged. can be a bigWig file or an R object.}

\item{are_R_objects}{Boolean term (true or false) to indicate if the inputted replicates are R objects. It'll use load() for the reps as opposed to reading them in via rtracklayer::import.bed(). Default to FALSE.}

\item{merged_samples_label}{Sample label of the merged bigWig file.}

\item{out_dir}{Output directory for merged bigWig files.}
}
\value{
a single merged bigWig file.
}
\description{
Merge normalized bigWig files, which are normally replicates. This function should be used after implementing the 'norm_bw()' function.
}
\details{
Title
}
