% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_norm_bw.R
\name{merge_norm_bw}
\alias{merge_norm_bw}
\title{Merging of normalized bigWig files.}
\usage{
merge_norm_bw(are_R_objects = FALSE, rep1, rep2, merged_samples_label, out_dir)
}
\arguments{
\item{are_R_objects}{a logical indicating whether the inputs are bigWig files or R objects.}

\item{rep1}{a character string specifying the first bigWig file or R object to be merged. Usually the first replicate.}

\item{rep2}{a character string specifying the second bigWig file or R object to be merged. Usually the second replicate.}

\item{merged_samples_label}{a character string specifying the filename of the merged bigWig file.}

\item{out_dir}{a character string specifying the output directory for the merged bigWig file.}
}
\value{
a single merged bigWig file.
}
\description{
Performs merging of normalized bigWig files, which are normally replicates.
}
\details{
Title
}
\examples{
merge_norm_bw(are_R_objects = TRUE,
              rep1 = system.file("extdata/data","PA_1.rda",
                                 package = "ChIPbinner"),
              rep2 = system.file("extdata/data","PA_2.rda",
                                 package = "ChIPbinner"),
              merged_samples_label = "merged_PA",
              out_dir = system.file("extdata",
                                    package = "ChIPbinner"))
}
