% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_bw.R
\name{norm_bw}
\alias{norm_bw}
\title{Title}
\usage{
norm_bw(
  wd,
  chromSizes,
  binned_files_path,
  treatment,
  control = NULL,
  chipLibrarySize,
  inputLibrarySize = NULL,
  cell_line,
  histone_mark,
  blacklist,
  addition,
  raw_count_cutoff,
  use_input,
  window_size,
  scaling_factor = NULL
)
}
\arguments{
\item{scaling_factor}{}
}
\description{
Title
}
