% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_bw.R
\name{norm_bw}
\alias{norm_bw}
\title{Normalization and conversion of binned BED files into bigWig.}
\usage{
norm_bw(
  out_dir,
  genome_assembly,
  use_input = TRUE,
  depth_norm = TRUE,
  immunoprecipitated_binned_file,
  input_binned_file = NULL,
  pseudocount = NULL,
  raw_count_cutoff = NULL,
  scaling_factor = NULL
)
}
\arguments{
\item{out_dir}{a character string specifying the output directory for the resulting bigWig files.}

\item{genome_assembly}{a character string specifying the genome assembly. Allowed values include "hg38" or "mm10".}

\item{use_input}{a logical indicating whether to use an input sample for normalization. This refers to the genomic input or IgG control sample for CUT&RUN/TAG.}

\item{depth_norm}{a logical indicating whether to normalize by library size. If the samples have previously been normalized by size factors, normalizing by library size is not needed.}

\item{immunoprecipitated_binned_file}{a character string specifying the binned BED file for the treated sample, normally the immunoprecipitated sample for ChIP-seq or the targeted, enriched sample for CUT&RUN/TAG.}

\item{input_binned_file}{a character string specifying the binned BED file for the control sample. This is usually the genomic input or IgG control sample for CUT&RUN/TAG.}

\item{pseudocount}{a numeric specifying the pseudocount to avoid division by 0. If NULL (the default), the pseudocount is 1e-3.}

\item{raw_count_cutoff}{a numeric specifying the raw read count cutoff to exclude bins depleted in signal across all tracks. For example, if 10 is inputted, this removes raw read count consistently lower than 10. If NULL (the default), the raw count cutoff is 0.}

\item{scaling_factor}{a numeric specifying the quantitative normalization/scaling of the raw binned signal by this factor. For example, this could be ChIP-Rx values or genome-wide modification percentage values obtained from mass spectrometry values. If NULL(the default), no quantitative scaling is performed.}
}
\value{
a single bigWig file with normalized binned scores.
}
\description{
Normalizes genome-wide binned scores and generates a bigWig file containing these normalized scores along with their corresponding genomic regions.
}
\details{
Title
}
\examples{
norm_bw(out_dir = system.file("extdata",package = "ChIPbinner"),
        genome_assembly = "hg38",
        immunoprecipitated_binned_file = system.file("extdata",
                                            "downsampled.Cal27.WT.H3K36me2.10kb.bed.gz",
                                            package = "ChIPbinner"),
        use_input = TRUE,
        depth_norm = TRUE,
        input_binned_file = system.file("extdata",
                                          "downsampled.Cal27.WT_input.H3K36me2.10kb.bed.gz",
                                          package = "ChIPbinner"),
        raw_count_cutoff = 0,
        pseudocount = 1e-3,
        scaling_factor = 0.450328805)
}
