% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_bw.R
\name{norm_bw}
\alias{norm_bw}
\title{Title}
\usage{
norm_bw(
  out_dir,
  chromSizes,
  blacklist,
  cell_line,
  histone_mark,
  treated_samp_label,
  treated_samp_file,
  treated_samp_library_size,
  control_label = NULL,
  control_file = NULL,
  control_library_size = NULL,
  use_input,
  window_size,
  addition,
  raw_count_cutoff,
  scaling_factor = NULL
)
}
\arguments{
\item{scaling_factor}{}
}
\description{
Title
}
