% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_clus.R
\name{pre_clus}
\alias{pre_clus}
\title{Title}
\usage{
pre_clus(
  wd,
  path_for_norm_bw,
  window_size,
  treatment,
  control,
  cell_line,
  histone_mark
)
}
\arguments{
\item{histone_mark}{}
}
\description{
Title
}
