% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_clust.R
\name{pre_clust}
\alias{pre_clust}
\title{Pre-processing the normalized bigWig files prior to generating clusters.}
\usage{
pre_clust(
  out_dir,
  treated_samp_norm_bw,
  wildtype_samp_norm_bw,
  output_filename,
  are_R_objects = FALSE
)
}
\arguments{
\item{out_dir}{Output directory for the matrix, BED file and preliminary scatterplot.}

\item{treated_samp_norm_bw}{The normalized bigwig file for the treated sample.}

\item{wildtype_samp_norm_bw}{The normalized bigwig file for the wildtype sample.}

\item{output_filename}{Name for the matrix file and pooled BED file.}

\item{are_R_objects}{Boolean term (true or false) to indicate if the inputted bigwig files are R objects. It'll use load() for the reps as opposed to reading them in via rtracklayer::import.bed(). Default to FALSE.}
}
\value{
a matrix file of scores for the two samples being compared, a pooled BED file of genomic coordinates for the bins and a preliminary scatterplot of the comparison.
}
\description{
This function pre-processes the bigWig files: it finds overlapping regions between two bigWig files, removes the bottom and top 1\% of bins across the two samples and finally generates a matrix of scores and BED file of coordinates to be used with HDBscan for generating clusters of bins with similar scores.
}
\details{
Title
}
\examples{
library(tidyverse)
library(rtracklayer)
library(GenomicRanges)
library(data.table)
library(lattice)
library(gridExtra)
library(DiffBind)
library(idr2d)
library(patchwork)
pre_clust(
  out_dir = system.file("extdata",package = "ChIPbinner"),
  treated_samp_norm_bw = system.file("extdata/data","downsampled.Cal27.WT.H3K36me2.10kb.norm_bw.rda", package = "ChIPbinner"),
  wildtype_samp_norm_bw = system.file("extdata/data","downsampled.Cal27.NSD1_KO.H3K36me2.10kb.norm_bw.rda", package = "ChIPbinner"),
  output_filename = "downsampled.Cal27.WT.NSD1_KO.H3K36me2.10kb",
  are_R_objects = TRUE
)
}
